### 2026-02-16 #10068 @arfon
@Alek050 – thanks for your submission to JOSS. Could you please ensure you have the following two sections in your paper please:

1. State of the Field: A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear "build vs. contribute" justification explaining your unique scholarly contribution and why existing alternatives are insufficient.a
2. Software Design: An explanation of the trade-offs you weighed, the design/architecture you chose, and why it matters for your research application. This should demonstrate meaningful design thinking beyond a superficial code structure description.
