### 2026-02-18 520cef5
First implementation Spearman pitch control
This is the first implementation of the pitch control suggested by Spearman(2017). It should work, but there are still a lot of things to be done, like mathematical refinement and also making the code more efficient
