+++ databallpy/features/pitch_control_spearman.py
@@ -1,0 +1,514 @@
+import re
+import numpy as np
+import pandas as pd
+
+def build_ndarray(
+        data: pd.Series, 
+        pattern: re.Pattern = None, 
+        type: str = "ball"
+        ) -> np.ndarray:
+    """
+    Takes data and puts it in certain form for later usage
+
+    Parameters
+    ------------
+    data: pd.Series
+    Series object representing a single frame of the tracking data
+
+    pattern: re.Pattern
+    Pattern object necessary for identifying the correct columns to put into the needed form
+
+    type: str
+    String object necessary to differentiate what type of data is supposed to be transformed
+
+    
+    Output
+    ----------
+    output: np.ndarray
+    Ndarray with the data reformed in the necessary shape
+
+    """
+
+
+    output = []
+
+    groups = {}
+
+    if type == "ball" or pattern == None:
+        pattern = re.compile(rf"ball_(x|y)")
+        x_val = None
+        y_val = None
+
+        for col in data.index:
+            match = pattern.match(col)
+            if match:
+                if match.group(1) == "x":
+                    x_val = data[col]
+                else:
+                    y_val = data[col]
+
+                if x_val is not None and y_val is not None:
+                    return np.asarray([[x_val, y_val]])
+
+
+    for col in data.index:
+        match = pattern.match(col)
+        if match:
+            idx = int(match.group(1))
+            xy = match.group(2)
+
+            if idx not in groups:
+                groups[idx] = {}
+
+            groups[idx][xy] = data[col]
+
+    if type == "position":
+        for idx in sorted(groups.keys()):
+            x_vals = groups[idx].get("x")
+            y_vals = groups[idx].get("y")
+            if np.isnan(x_vals):
+                continue
+            output.append((x_vals, y_vals))
+    
+    elif type == "velocity":
+        for idx in sorted(groups.keys()):
+            x_vals = groups[idx].get("vx")
+            y_vals = groups[idx].get("vy")
+            if np.isnan(x_vals):
+                continue
+            output.append((x_vals, y_vals))
+
+
+    return np.asarray(output)
+
+def prepare_data(
+        data: pd.Series
+        ) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Function to prepare all needed data from a single frame
+
+    Parameters
+    ------------
+    data: pd.Series
+    Series object representing a single frame in the tracking data
+
+
+    Output
+    ---------
+    home_pos: np.ndarray
+    Positions of the home players in the form of an array with the dimensions (n, 2)
+
+    away_pos: np.ndarray
+    Positions of the away players in the form of an array with the dimensions (n, 2)
+
+    home_vel: np.ndarray
+    Velocities of the home players in the form of an array with the dimensions (n, 2)
+
+    away_vel: np.ndarray
+    Velocities of the away players in the form of an array with the dimensions (n, 2)
+    """
+
+    ball_pos = build_ndarray(data)
+
+    pattern_home_pos = re.compile(rf"home_(\d+)_(x|y)$")
+    home_pos = build_ndarray(data, pattern_home_pos, "position")
+
+    pattern_away_pos = re.compile(rf"away_(\d+)_(x|y)$")
+    away_pos = build_ndarray(data, pattern_away_pos, "position")
+
+    pattern_home_vel = re.compile(rf"home_(\d+)_(vx|vy)$")
+    home_vel = build_ndarray(data, pattern_home_vel, "velocity")
+    
+    pattern_away_vel = re.compile(rf"away_(\d+)_(vx|vy)$")
+    away_vel = build_ndarray(data, pattern_away_vel, "velocity")
+
+    return home_pos, away_pos, home_vel, away_vel, ball_pos
+
[…]
