### 2026-02-18 89e446e
Added platformdirs & python3.14
