+++ pyproject.toml
@@ -23,25 +23,27 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "License :: OSI Approved :: MIT License",
     "Topic :: Scientific/Engineering",
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.14"
-numpy = "^2.1"
+python = ">=3.10,<3.15"
+numpy = ">=2.1,<3.0"
 beautifulsoup4 = "^4.11.1"
 lxml = ">=4.9,<=6.0"
 tqdm = "^4.64.1"
 matplotlib = "^3.6.3"
 requests = "^2.31"
 scipy = "^1.11"
-pandas = "^2.1"
+pandas = ">=2.1,<3.0"
 chardet = "^5.2.0"
 pandera = "^0.22.1"
 pyarrow = "^19.0.0"
 kloppy = { version = ">=3.18.0", optional = true }
 accessible-space = { version = ">=2.0.13", optional = true }
+platformdirs = "^4.9.2"
 
 [tool.poetry.extras]
 kloppy = ["kloppy"]
