+++ databallpy/schemas/players.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-import pandera as pa
+import pandera.pandas as pa
 
 from databallpy.utils.constants import DATABALLPY_POSITIONS
 
