+++ pyproject.toml
@@ -36,10 +36,13 @@ lxml = ">=4.9,<=6.0"
 tqdm = "^4.64.1"
 matplotlib = "^3.6.3"
 requests = "^2.31"
-scipy = "^1.11"
+scipy =[
+  { version = ">=1.10,<2.0", python = ">=3.10,<3.14" },
+  { version = ">=1.16.1,<2.0", python = ">=3.14,<4.0" }
+]
 pandas = ">=2.1,<3.0"
 chardet = "^5.2.0"
-pandera = "^0.22.1"
+pandera = "^0.29.0"
 pyarrow = "^19.0.0"
 kloppy = { version = ">=3.18.0", optional = true }
 accessible-space = { version = ">=2.0.13", optional = true }
