+++ tests/features/test_filters.py
@@ -5,7 +5,14 @@ from unittest.mock import patch
 import numpy as np
 import pandas as pd
 
-from databallpy.features.filters import _filter_data, filter_tracking_data
+from databallpy.features.filters import _filter_data, filter_tracking_data, _savgol_with_nan_compat
+import warnings
+
+import pytest
+from numpy.testing import assert_allclose
+
+from scipy.signal import savgol_filter
+
 
 
 class TestFilters(unittest.TestCase):
@@ -46,7 +53,7 @@ class TestFilters(unittest.TestCase):
     def test_filter_data_sg(self):
         # Test that filter_data returns the expected output for a Savitzky-Golay filter
         arr = np.array([1, 3, 3, 4, 5])
-        expected_output = np.array([1.333333333, 2.3333333333, 3.33333333, 4.0, 5.0])
+        expected_output = np.array([1.33, 2.33, 3.33, 4.0, 5.0])
         self.assertTrue(
             np.allclose(
                 _filter_data(
@@ -172,11 +179,108 @@ class TestFilters(unittest.TestCase):
             {
                 "home_1_x": [10, 20, -30, 40, np.nan, 60],
                 "home_1_y": [5, 12, -20, 30, np.nan, 60],
-                "ball_x": [20, 0, 10, np.nan, np.nan, np.nan],
-                "ball_y": [11.5, -1, 7.333333, np.nan, np.nan, np.nan],
+                "ball_x": [20, 0, 10, 20., np.nan, 60.],
+                "ball_y": [11.5, -1, 7.33, 18.33, np.nan, 60.],
                 "away_13_x": [10, 20, -30, 40, np.nan, 60],
                 "away_13_y": [5, 12, -20, 30, np.nan, 60],
             }
         )
 
         pd.testing.assert_frame_equal(filtered_data, expected_output)
+
+
+
+class TestSavgolWithNanCompat:
+    def test_no_nans_behaves_like_plain_savgol(self) -> None:
+        """When there are no NaNs, the wrapper should behave like savgol_filter.
+
+        It should:
+        - not modify NaNs (there are none),
+        - not emit warnings,
+        - produce the same output (up to the .round(2) in the implementation).
+        """
+        rng = np.random.default_rng(42)
+        arr = rng.normal(size=20)
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter("error")  # fail on unexpected warnings
+            result = _savgol_with_nan_compat(arr, window_length=5, polyorder=2)
+
+        assert len(w) == 0, "No warnings expected when there are no NaNs"
+
+        expected = savgol_filter(arr.astype(float), 5, 2, mode="interp").round(2)
+        assert result.shape == arr.shape
+        assert_allclose(result, expected, rtol=0, atol=0)  # exact after rounding
+
+    def test_with_nans_preserves_nan_positions(self) -> None:
+        """NaNs should be preserved in their original positions after filtering.
+
+        Finite values should be smoothed; NaN positions remain NaN.
+        """
+        arr = np.array([1.0, -1.0, np.nan, 4.0, 5.0, np.nan, 7.0], dtype=float)
+
+        result = _savgol_with_nan_compat(arr, window_length=5, polyorder=2)
+
+        assert result.shape == arr.shape
+
+        # NaN mask should be identical
+        assert np.array_equal(np.isnan(result), np.isnan(arr)), \
+            "NaN positions must be preserved"
+
+        # At least one finite value should be changed by the smoothing
+        finite_mask = np.isfinite(arr)
+        assert np.any(result[finite_mask] != arr[finite_mask])
+
+    def test_with_nans_does_not_raise_or_warn_when_enough_valid_samples(self) -> None:
[…]
