+++ pyproject.toml
@@ -43,7 +43,10 @@ scipy =[
 pandas = ">=2.1,<3.0"
 chardet = "^5.2.0"
 pandera = "^0.29.0"
-pyarrow = "^19.0.0"
+pyarrow = [
+  { version = ">=19.0.0,<23.0.0", python = ">=3.10,<3.14" },
+  { version = ">=23.0.0,<100.0.0", python = ">=3.14" }
+]
 kloppy = { version = ">=3.18.0", optional = true }
 accessible-space = { version = ">=2.0.13", optional = true }
 platformdirs = "^4.9.2"
@@ -75,7 +78,10 @@ jupyter-book = "^1.0.0"
 ipykernel = "^6.29.5"
 
 [tool.poetry.group.profiler.dependencies]
-line-profiler = "^4.1.2"
+line-profiler = [
+  { version = ">=4.1.0,<5.0.0", python = "<3.14" },
+  { version = ">=5.0.1,<6.0.0", python = "=3.14"}
+]
 
 [tool.semantic_release]
 version_variable = "pyproject.toml:version" # version location
