+++ pyproject.toml
@@ -32,7 +32,6 @@ classifiers = [
 python = ">=3.10,<3.15"
 numpy = ">=2.1,<3.0"
 beautifulsoup4 = "^4.11.1"
-lxml = ">=4.9,<=6.0"
 tqdm = "^4.64.1"
 matplotlib = "^3.6.3"
 requests = "^2.31"
