### 2026-02-18 559b0b6
changed np version for python3.14 for windows
