+++ pyproject.toml
@@ -30,7 +30,11 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.15"
-numpy = ">=2.1,<3.0"
+numpy = [
+    { version = ">=1.26.4,<2.3.0", python = ">=3.10,<3.14" },
+    { version = ">=2.3.2,<3.0.0", python = ">=3.14" }
+]
+
 beautifulsoup4 = "^4.11.1"
 tqdm = "^4.64.1"
 matplotlib = "^3.6.3"
