### 2026-02-18 e1e0b47
Added platformdirs & python3.14 (#347)
* Added platformdirs & python3.14

* retry

* retry v2

* fixed linting

* removed lxml

* linters

* changed np version for python3.14 for windows
