+++ tests/test_game.py
@@ -3,7 +3,7 @@ import unittest
 
 import numpy as np
 import pandas as pd
-import pandera as pa
+import pandera.pandas as pa
 
 from databallpy.game import Game
 from databallpy.schemas import EventData, TrackingData
