+++ databallpy/data_parsers/event_data_parsers/fifa_parser.py
@@ -8,32 +8,28 @@ from databallpy.events import (
     IndividualCloseToBallEvent,
     PassEvent,
     ShotEvent,
-    TackleEvent,
 )
 from databallpy.utils.constants import MISSING_INT
 from databallpy.utils.logging import logging_wrapper
+from databallpy.utils.tz_modification import utc_to_local_datetime
 
-# FIFA event type mappings to databallpy events
 FIFA_TO_DATABALLPY_MAP = {
     "pass": "pass",
     "assist": "pass",
     "cross": "pass",
     "attempt_at_goal": "shot",
+    "goal": "shot",     
     "own_goal": "own_goal",
-    "tackle": "tackle",
-    "no_event": "tackle",
 }
 
-# Shot outcome mappings
 SHOT_OUTCOMES = {
     "incomplete": "miss_off_target",
     "off_target": "miss_off_target",
-    "on_target": "on_target",
+    "on_target": "miss_on_target",
     "complete": "goal",
     "own_goal": "own_goal",
 }
 
-# Body part mappings
 BODY_PART_MAP = {
     "right_foot": "right_foot",
     "left_foot": "left_foot",
@@ -43,7 +39,6 @@ BODY_PART_MAP = {
     "feet": "other",
 }
 
-# Set piece mappings
 SET_PIECE_MAP = {
     "freekick": "free_kick",
     "throwin": "throw_in",
@@ -53,18 +48,18 @@ SET_PIECE_MAP = {
     "penalty": "penalty",
 }
 
-# Phase of the game mappings
-PHASE_TYPE_MAP = {
-    "referee_action": "referee_action",
-    "in_possession": "in_possession",
-    "out_of_possession": "out_of_possession",
-    "in_contest": "in_contest",
+FIFA_LINE_BREAK_TO_PASS_TYPE = {
+    "over": "long_ball",
+    "through": "through_ball",
+    "around": "through_ball",
 }
 
 
 @logging_wrapper(__file__)
 def load_fifa_event_data(
-    metadata_loc: str, events_loc: str, pitch_dimensions: list = [105.0, 68.0]
+    metadata_loc: str,
+    events_loc: str,
+    pitch_dimensions: list = [105.0, 68.0],
 ) -> tuple[
     pd.DataFrame, Metadata, dict[str, dict[str | int, IndividualCloseToBallEvent]]
 ]:
@@ -77,7 +72,7 @@ def load_fifa_event_data(
     Args:
         metadata_loc (str): location of the metadata JSON file.
         events_loc (str): location of the events JSON file.
-        pitch_dimensions (list, optional): the length and width of the pitch in meters
+        pitch_dimensions (list, optional): the length and width of the pitch in meters.
 
     Returns:
         Tuple[pd.DataFrame, Metadata, dict]: the event data of the game, the metadata,
@@ -87,11 +82,11 @@ def load_fifa_event_data(
         raise TypeError(f"metadata_loc should be a string, not a {type(metadata_loc)}")
     if not isinstance(events_loc, str):
         raise TypeError(f"events_loc should be a string, not a {type(events_loc)}")
-    if not metadata_loc[-5:] == ".json":
+    if not metadata_loc.endswith(".json"):
         raise ValueError(
             f"metadata file should be of .json format, not {metadata_loc.split('.')[-1]}"
         )
-    if not events_loc[-5:] == ".json":
+    if not events_loc.endswith(".json"):
         raise ValueError(
             f"events file should be of .json format, not {events_loc.split('.')[-1]}"
         )
@@ -100,7 +95,11 @@ def load_fifa_event_data(
     all_players = pd.concat(
[…]
