+++ tests/data_parsers/event_data_parsers/test_fifa_parser.py
@@ -1,0 +1,580 @@
+"""
+Unit tests for fifa_parser.py, following the style of test_opta_parser.py.
+
+Run with:  pytest test_fifa_parser.py -v
+"""
+
+import os
+import unittest
+
+import numpy as np
+import pandas as pd
+
+from databallpy.data_parsers import Metadata
+from databallpy.events import PassEvent, ShotEvent
+from databallpy.utils.constants import MISSING_INT
+from fifa_parser import (
+    _determine_period_flips,
+    _get_game_score,
+    _get_player_info,
+    _get_transformed_coordinates,
+    _load_event_data,
+    _load_metadata,
+    _make_pass_instance,
+    _make_shot_event_instance,
+    load_fifa_event_data,
+)
+
+TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "test_data")
+METADATA_LOC = os.path.join(TEST_DATA_DIR, "fifa_metadata_test.json")
+EVENTS_LOC = os.path.join(TEST_DATA_DIR, "fifa_events_test.json")
+
+_UTC = "UTC"
+_CET = "Europe/Amsterdam"
+_KO_UTC = pd.Timestamp("2023-01-01T14:00:00", tz=_UTC)
+
+
+def _evt_dt(ms: int) -> pd.Timestamp:
+    """Event datetime in CET given match_time_in_ms."""
+    return (_KO_UTC + pd.to_timedelta(ms / 1000.0, unit="s")).tz_convert(_CET)
+
+
+_PERIODS_FRAMES = pd.DataFrame(
+    {
+        "period_id": [1, 2, 3, 4, 5],
+        "start_datetime_ed": pd.to_datetime(
+            ["2023-01-01T14:00:00Z", "2023-01-01T15:00:00Z", pd.NaT, pd.NaT, pd.NaT],
+            utc=True,
+        ).tz_convert(_CET),
+        "end_datetime_ed": pd.to_datetime(
+            ["2023-01-01T14:45:00Z", "2023-01-01T15:45:00Z", pd.NaT, pd.NaT, pd.NaT],
+            utc=True,
+        ).tz_convert(_CET),
+    }
+)
+
+HOME_PLAYERS_FIFA = pd.DataFrame(
+    {
+        "id": [1001, 1002],
+        "full_name": ["HOME KEEPER", "HOME STRIKER"],
+        "formation_place": [MISSING_INT, MISSING_INT],
+        "position": ["unspecified", "unspecified"],
+        "starter": [False, False],
+        "shirt_num": [1, 9],
+    }
+)
+
+AWAY_PLAYERS_FIFA = pd.DataFrame(
+    {
+        "id": [2001, 2002],
+        "full_name": ["AWAY KEEPER", "AWAY STRIKER"],
+        "formation_place": [MISSING_INT, MISSING_INT],
+        "position": ["unspecified", "unspecified"],
+        "starter": [False, False],
+        "shirt_num": [1, 9],
+    }
+)
+
+MD_FIFA = Metadata(
+    game_id=99999,
+    pitch_dimensions=[105.0, 68.0],
+    periods_frames=_PERIODS_FRAMES,
+    frame_rate=MISSING_INT,
+    home_team_id=100,
+    home_team_name="HOME TEAM",
+    home_players=HOME_PLAYERS_FIFA,
+    home_score=1,
+    home_formation="4231",
+    away_team_id=200,
+    away_team_name="AWAY TEAM",
+    away_players=AWAY_PLAYERS_FIFA,
+    away_score=2,
+    away_formation="442",
+    country="Netherlands",
+)
+
[…]

[…]
