### 2026-02-23 4a7a535
Added State of the field and Software Design subsections, updated some DOI's
