+++ joss-paper/paper.bib
@@ -182,7 +182,7 @@
 @article{Hader2019,
    abstract = {Background: Monitoring athletes’ external load during a soccer match may be useful to predict post-match acute and residual fatigue. This estimation would allow individual adjustments to training programs to minimize injury risk, improve well-being, and restore players’ physical performance and inform the recovery process. Methods: Using a systematic review and meta-analysis of the literature, the aim is to determine which monitoring variables would be the strongest predictors of acute (immediately) and residual (up to 72 h) fatigue states in soccer. PubMed, SPORTDiscus, and Web of Science databases were searched (until September 2018). Studies concurrently examining soccer match-related external load metrics and subjective and/or objective measures were selected to determine pooled correlations (r¯) with confidence intervals (CI). The quality and strength of the findings of each study were evaluated to identify overall levels of evidence. Results: Eleven studies were included (n = 165 athletes). Acute (r¯ = 0.67; 95% CI = [0.40, 0.94]) and residual (24 h post-match, r¯ = 0.54; 95% CI = [0.35, 0.65]) changes in muscle damage markers and countermovement jump peak power output (CMJPPO) were, with moderate to strong evidence, largely correlated with running distance above 5.5 m s−1. No other external load metric was largely correlated with both biochemical and neuromuscular markers. For every 100-m run above 5.5 m·s−1, CK activity measured 24 h post-match increased by 30% and CMJPPO decreased by 0.5%. Conversely, the total distance covered did not present any evidence of a clear relationship with any fatigue-related marker at any time-point. Conclusions: Running distance above 5.5 m·s−1 represents the most sensitive monitoring variable characterizing biochemical and neuromuscular responses, at least when assessed during the initial 24 h (not at 48 h/72 h) post-match recovery period. In addition, total distance covered is not sensitive enough to inform decision-making during the fatigue monitoring process.},
    author = {Karim Hader and Michael C. Rumpf and Maxime Hertzog and Liam P. Kilduff and Olivier Girard and Joao R. Silva},
-   doi = {10.1186/S40798-019-0219-7/FIGURES/2},
+   doi = {10.1186/S40798-019-0219-7},
    issn = {21989761},
    issue = {1},
    journal = {Sports Medicine - Open},
@@ -334,7 +334,7 @@ journal = {Journal of Open Source Software}
 @article{Robertson2023,
    abstract = {Electronic performance & tracking systems (EPTS) are commonly used to track the location and velocity of athletes in many team sports. A range of associated applications using the derived data exist, such as assessment of athlete characteristics, informing training design, assisting match adjudication and providing fan insights for broadcast. Consequently the quality of such systems is of importance to a range of stakeholders. The influence of both systematic and methodological factors such as hardware, software settings, sample rate and filtering on this resulting quality is non-trivial. Highlighting these allows for the user to understand their strengths and limitations in various decision-making processes, as well as identify areas for research and development. In this paper, a number of challenges and considerations relating to the determination of EPTS validity for team sport are outlined and discussed. The aim of this paper is to draw attention of these factors to both researchers and practitioners looking to inform their decision-making in the EPTS area. Addressing some of the posited considerations in future work may represent best practice; others may require further investigation, have multiple potential solutions or currently be intractable.},
    author = {Sam Robertson and Grant Malcolm Duthie and Kevin Ball and Bartholomew Spencer and Fabio Rubens Serpiello and Jade Haycraft and Nicolas Evans and Johsan Billingham and Robert James Aughey},
-   doi = {10.3389/FSPOR.2023.1266522/BIBTEX},
+   doi = {10.3389/fspor.2023.1266522},
    issn = {26249367},
    journal = {Frontiers in Sports and Active Living},
[…]
