+++ joss-paper/paper.md
@@ -33,31 +33,35 @@ date: 05-11-2025
 bibliography: paper.bib
 ---
 
+
 # Summary
 
-Over the last decade, there has been a growing interest in soccer analytics from different backgrounds and for different use cases. Examplary use cases are: first, practical decision making and benchmarking of players based on aggregated metrics such as pass success percentage and expected goals (xG) [@Goes2020a]. Second, using internal and external load metrics for training periodization and injury predictions [@Hader2019]. Third, basic behavioural science soccer with a focus on group and subgroup behaviour[@Goes2020b]. The interest in soccer analysis has also increased since data has become more openly available [@Bassek2025]. However, a key challenge is that every data provider uses their own data format, which makes it hard to compare and switch between different providers and create large datasets that encompass different leagues and competitions. Currently, open-source packages like [Kloppy](https://kloppy.pysport.org) try to overcome this challenge by providing a uniform data format. Similarly, the scientific side proposes a common data format for soccer game data [@Anzer2025]. While Kloppy focuses primarily on parsing soccer data, Floodlight [@Raabe2022] delivers a framework for physical analysis of team sports, and [mplsoccer](https://github.com/andrewRowlinson/mplsoccer) is widely utilized for visualising soccer data. 
+Over the last decade, there has been a growing interest in soccer analytics from different backgrounds and for different use cases. Example use cases are: first, practical decision making and benchmarking of players based on aggregated metrics such as pass success percentage and expected goals (xG) [@Goes2020a]. Second, using internal and external load metrics for training periodization and injury predictions [@Hader2019]. Third, basic behavioural science soccer with a focus on group and subgroup behaviour[@Goes2020b]. The interest in soccer analysis has also increased since data has become more openly available [@Bassek2025]. However, a key challenge is that every data provider uses their own data format, which makes it hard to compare and switch between different providers and create large datasets that encompass different leagues and competitions. Currently, open-source packages like [Kloppy](https://kloppy.pysport.org) try to overcome this challenge by providing a uniform data format. Similarly, the scientific side proposes a common data format for soccer game data [@Anzer2025]. While Kloppy focuses primarily on parsing soccer data, Floodlight [@Raabe2022] delivers a framework for physical analysis of team sports, and [mplsoccer](https://github.com/andrewRowlinson/mplsoccer) is widely utilized for visualising soccer data. 
 
 Lately, there has been a growing interest in combining event and tracking data for contextualised tactical analysis of soccer games. This provides the possibility to not only know that a pass happened at a specific moment in the match (event data) but also what the defensive structure was during this pass [@Forcher2022; @Herold2022], and what other passing options were available at this moment (tracking data) [@Spearman2017]. Contextual analysis goes beyond aggregated metrics and provides the ability to do quantitative analysis of single moments or specific phases in the game [@Oonk2025a; @Jerome2024]. Merging tracking and event data is a key challenge for contextualised analysis of soccer games. [`DataBallPy`](https://databallpy.readthedocs.io/en/latest/) is an open source python package for contextual analysis of soccer games because (1) it uses a standardized data format for both event and tracking data, (2) it provides a framework where all data of a game is bundled, instead of considered as seperate data objects, (3) it includes a high quality and learning free synchronsiation algorithm that works on any combination of tracking and event data providers, and (4) it has integrated multiple practical and scientific features within the package that allow for efficient computation with minimal user input. 
 
 
 # Statement of need
 
[…]
