+++ joss-paper/paper.md
@@ -47,13 +47,13 @@ Modern soccer analytics increasingly rely on both event data and tracking data f
 
 This added context provides insights and nuances, primarily on a tactical level, that neither event nor tracking data can provide independently. For example, shot events are enriched with information about defensive and keeper positioning to create better expected goals models [@Anzer2021], passes are evaluated by making risk reward assessments of all possible passing options [@Goes2021], determinants of successful 1v1 actions are modelled from spatiotemporal features [@Oonk2025a], and the spatiotemporal context of events is used to predict dangerousity of a game state [@Fernandez2021].
 
-A contextual analysis requires a proper synchronisation of event and tracking data. Although event and tracking data often both provide timestamps, their alignment has shown to be extremely poor with reported errors of 1.82 (+-4.06) seconds [@Anzer2021]. Especially, the random error is concerning since it does not allow for easy correction, and within 4 seconds, the game might have evolved to an entirely different situation, which impacts the stability of the found effects. `@Oonk2025b` showed that the expected goals model decreased in Brier loss from 0.096 to 0.082 (lower is better) when using the synchronised data compared to a naive timestamp alignment. Similarly, the feature importance of features that relied on combined tracking and event data information was close to 0 in the timestamp synchronisation model, which was not the case for the properly synchronised situation [@Oonk2025b]. Therefore, there is a need for software that is easy to use, has a state-of-the-art synchronisation algorithm, and provides a convenient data structure for any subsequent analysis.
+A contextual analysis requires a proper synchronisation of event and tracking data. Although event and tracking data often both provide timestamps, their alignment has shown to be extremely poor with reported errors of 1.82 (+-4.06) seconds [@Anzer2021]. Especially, the random error is concerning since it does not allow for easy correction, and within 4 seconds, the game might have evolved to an entirely different situation, which impacts the stability of the found effects. @Oonk2025b showed that the expected goals model decreased in Brier loss from 0.096 to 0.082 (lower is better) when using the synchronised data compared to a naive timestamp alignment. Similarly, the feature importance of features that relied on combined tracking and event data information was close to 0 in the timestamp synchronisation model, which was not the case for the properly synchronised situation [@Oonk2025b]. Therefore, there is a need for software that is easy to use, has a state-of-the-art synchronisation algorithm, and provides a convenient data structure for any subsequent analysis.
 
 # State of the field
 
 The currently available packages allow for the parsing and analysis of tracking and event data independently.  ([Kloppy](https://kloppy.pysport.org)) is a well-known data parsing package in the soccer analytics field. Its primary focus is to simplify and standardize parsing soccer tracking and event data from different data providers. A similar project aims to enforce a standardized format for soccer, which could make different parsers redundant in the future [@Anzer2025]. Other packages allow for analysis [@Raabe2022] and visualisation ([mplsoccer](https://github.com/andrewRowlinson/mplsoccer)) of soccer tracking and event data. 
 
[…]
