+++ joss-paper/paper.bib
@@ -241,6 +241,13 @@
       primaryClass={cs.DB},
       url={https://arxiv.org/abs/2508.09238}, 
 }
+
+@misc{Kwiakowski
+   author = {M Kwiakowski and A Clark},
+   title = {The right way to synchronise event and tracking data},
+   url = {https://kwiatkowski.io/sync.soccer}
+}
+
 @article{Linke2020,
    abstract = {The present study aimed to validate and compare the football-specific measurement accuracy of two optical tracking systems engineered by TRACAB. The “Gen4” system consists of two multi-camera units (a stereo pair) in two locations either side of the halfway line, whereas the distributed “Gen5” system combines two stereo pairs on each side of the field as well as two monocular systems behind the goal areas. Data were collected from 20 male football players in two different exercises (a football sport-specific running course and small-sided games) in a professional football stadium. For evaluating the accuracy of the systems, measures were compared against simultaneously recorded measures of a reference system (VICON motion capture system). Statistical analysis uses RMSE for kinematic variables (position, speed and acceleration) and the difference in percentages for performance indicators (e.g. distance covered, peak speed) per run compared to the reference system. Frames in which players were obviously not tracked were excluded. Gen5 had marginally better accuracy (0.08 m RMSE) for position measurements than Gen4 (0.09 m RMSE) compared to the reference. Accuracy difference in instantaneous speed (Gen4: 0.09 m⋅s-1 RMSE; Gen5: 0.08 m⋅s-1 RMSE) and acceleration (Gen4: 0.26 m⋅s-2 RMSE; Gen5: 0.21 m⋅s-2 RMSE) measurements were significant, but also trivial in terms of the effect size. For total distance travelled, both Gen4 (0.42 ± 0.60%) and Gen5 (0.27 ± 0.35%) showed only trivial deviations compared to the reference. Gen4 showed moderate differences in the low-speed distance travelled category (-19.41 ± 13.24%) and small differences in the high-speed distance travelled category (8.94 ± 9.49%). Differences in peak speed, acceleration and deceleration were trivial (<0.5%) for both Gen4 and Gen5. These findings suggest that Gen5’s distributed camera architecture has minor benefits over Gen4’s single-view camera architecture in terms of accuracy. We assume that the main benefit of the Gen5 towards Gen4 lies in increased robustness of the tracking when it comes to optical overlapping of players. Since differences towards the reference system were very low, both TRACAB’s tracking systems can be considered as valid technologies for football-specific performance analyses in the settings tested as long as players are tracked correctly.},
    author = {Daniel Linke and Daniel Link and Martin Lames},
