+++ joss-paper/paper.bib
@@ -242,7 +242,7 @@
       url={https://arxiv.org/abs/2508.09238}, 
 }
 
-@misc{Kwiakowski
+@misc{Kwiakowski,
    author = {M Kwiakowski and A Clark},
    title = {The right way to synchronise event and tracking data},
    url = {https://kwiatkowski.io/sync.soccer}
