### 2026-02-23 #10068 @Alek050
Hi @arfon,

Thanks for you for opening the pre-review. I added the requested sections to the paper.md.

Regarding potential reviewers, these are two projects that are covered in JOSS that I have used before and have some similarity with `DataBallPy`:
- the KineticsToolkit (in your proposed list)
- [Floodlight](https://joss.theoj.org/papers/10.21105/joss.04588)

Let me know if there is anything else I can do to help.
