+++ tests/data_parsers/event_data_parsers/test_fifa_parser.py
@@ -9,11 +9,7 @@ import unittest
 
 import numpy as np
 import pandas as pd
-
-from databallpy.data_parsers import Metadata
-from databallpy.events import PassEvent, ShotEvent
-from databallpy.utils.constants import MISSING_INT
-from fifa_parser import (
+from databallpy.data_parsers.event_data_parsers.fifa_parser import (
     _determine_period_flips,
     _get_game_score,
     _get_player_info,
@@ -24,6 +20,10 @@ from fifa_parser import (
     _make_shot_event_instance,
     load_fifa_event_data,
 )
+
+from databallpy.data_parsers import Metadata
+from databallpy.events import PassEvent, ShotEvent
+from databallpy.utils.constants import MISSING_INT
 
 TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), "test_data")
 METADATA_LOC = os.path.join(TEST_DATA_DIR, "fifa_metadata_test.json")
@@ -102,8 +102,13 @@ ED_FIFA = pd.DataFrame(
         "seconds": [0.0, 5.0, 30.0, 0.0, 40.0, 20.0, 0.0],
         "player_id": [1002, 1002, 2002, 2002, 1002, 2002, 1001],
         "player_name": [
-            "HOME STRIKER", "HOME STRIKER", "AWAY STRIKER", "AWAY STRIKER",
-            "HOME STRIKER", "AWAY STRIKER", "HOME KEEPER",
+            "HOME STRIKER",
+            "HOME STRIKER",
+            "AWAY STRIKER",
+            "AWAY STRIKER",
+            "HOME STRIKER",
+            "AWAY STRIKER",
+            "HOME KEEPER",
         ],
         "team_id": [100, 100, 200, 200, 100, 200, 100],
         "is_successful": pd.array(
@@ -123,12 +128,22 @@ ED_FIFA = pd.DataFrame(
         "phase": ["in_possession"] * 7,
         "original_event_id": [101, 102, 103, 104, 105, 106, 107],
         "original_event": [
-            "kickoff", "pass", "cross", "attempt_at_goal",
-            "goal", "goal", "own_goal",
+            "kickoff",
+            "pass",
+            "cross",
+            "attempt_at_goal",
+            "goal",
+            "goal",
+            "own_goal",
         ],
         "original_event_type": [
-            "start_restart", "pass_like", "pass_like", "shot",
-            "shot", "shot", "shot",
+            "start_restart",
+            "pass_like",
+            "pass_like",
+            "shot",
+            "shot",
+            "shot",
+            "shot",
         ],
     }
 )
@@ -339,155 +354,302 @@ class TestFifaParser(unittest.TestCase):
     def test_make_pass_instance(self):
         """Regular pass: right_foot, possession_complete, no line_break."""
         event = {
-            "team_id": 100, "from_player_id": 1002, "event_id": 102,
-            "event_type": "pass_like", "category": "in_possession",
-            "event": "pass", "half_time": 1, "match_time_in_ms": 5000,
-            "side": None, "x": 0.25, "y": 0.5,
-            "outcome": "possession_complete", "body_type": "right_foot",
-            "origin": None, "line_break_direction": None,
-            "x_location_start": 0.25, "y_location_start": 0.5,
-            "x_location_end": 0.75, "y_location_end": 0.5,
+            "team_id": 100,
+            "from_player_id": 1002,
+            "event_id": 102,
+            "event_type": "pass_like",
+            "category": "in_possession",
+            "event": "pass",
+            "half_time": 1,
+            "match_time_in_ms": 5000,
+            "side": None,
+            "x": 0.25,
+            "y": 0.5,
+            "outcome": "possession_complete",
+            "body_type": "right_foot",
[…]
