### 2026-02-24 328f7a7
fix for pandera & pandas integration python 3.14
