### 2026-02-24 4e25a58
fix for pandera & pandas integration python 3.14 (#348)
