+++ tests/test_game.py
@@ -3,8 +3,11 @@ import unittest
 
 import numpy as np
 import pandas as pd
-import pandera.pandas as pa
 
+try:
+    import pandera.pandas as pa
+except ModuleNotFoundError:
+    import pandera as pa
 from databallpy.game import Game
 from databallpy.schemas import EventData, TrackingData
 from databallpy.utils.errors import DataBallPyError
