+++ databallpy/schemas/event_data.py
@@ -1,7 +1,11 @@
 from typing import Optional
 
 import pandas as pd
-import pandera.pandas as pa
+
+try:
+    import pandera.pandas as pa
+except ModuleNotFoundError:
+    import pandera as pa
 
 from databallpy.utils.constants import DATABALLPY_EVENTS
 
