### 2026-03-02 #10068 @arfon
@felixchenier @draabe :wave: would you be willing to review this submission for JOSS? The submission under consideration is _DataBallPy: Load, Synchronise, and Analyse your Soccer Data_

Based on your experience, and previously published papers here, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
