### 2026-03-05 #10068 @arfon
@alcantarar, @melund, @gagolews, @kanishkan91 👋 would you be willing to review this submission for JOSS? The submission under consideration is _DataBallPy: Load, Synchronise, and Analyse your Soccer Data_

Based on your experience, and previously reviewing experience for JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
