+++ tests/schemas/test_event_data_to_xml.py
@@ -1,0 +1,260 @@
+import tempfile
+import os
+import unittest
+import xml.etree.ElementTree as ET
+
+import pandas as pd
+
+from databallpy import Event, LabelDict, events_to_xml
+from databallpy.schemas.event_data import EventData
+
+
+def make_event_data():
+    """Minimal 4-row EventData spanning 2 periods."""
+    data = {
+        "event_id": [1, 2, 3, 4],
+        "databallpy_event": ["pass", "shot", "pass", None],
+        "period_id": [1, 1, 2, 2],
+        "minutes": [2, 44, 47, 89],
+        "seconds": [30.0, 0.0, 5.0, 55.0],
+        "player_id": [10, 20, 10, 20],
+        "player_name": ["Jan", "Piet", "Jan", "Piet"],
+        "team_id": [1, 2, 1, 2],
+        "team_name": ["Ajax", "PSV", "Ajax", "PSV"],
+        "is_successful": pd.array([True, False, True, None], dtype=pd.BooleanDtype()),
+        "start_x": [0.0, 10.0, -5.0, 20.0],
+        "start_y": [0.0, 5.0, -3.0, 10.0],
+        "datetime": [None, None, None, None],
+        "original_event_id": [100, 200, 300, 400],
+        "original_event": ["pass", "shot", "pass", "tackle"],
+    }
+    return EventData(data)
+
+
+class TestEventDataToXml(unittest.TestCase):
+    def setUp(self):
+        self.ed = make_event_data()
+
+    def _parse(self, xml_str):
+        # Strip the XML declaration before parsing
+        if xml_str.startswith("<?xml"):
+            xml_str = xml_str[xml_str.index("?>") + 2:].lstrip()
+        return ET.fromstring(xml_str)
+
+    def _instance_codes(self, root):
+        return [el.text for el in root.findall(".//instance/code")]
+
+    def test_to_xml_no_filters(self):
+        xml_str = self.ed.to_video_analysis_xml()
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
+        # 4 events + 2 period markers
+        self.assertEqual(len(instances), 6)
+        codes = self._instance_codes(root)
+        self.assertIn("1H", codes)
+        self.assertIn("2H", codes)
+
+    def test_to_xml_team_filter(self):
+        xml_str = self.ed.to_video_analysis_xml(team_id=1, tag_period_starts=False)
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
+        self.assertEqual(len(instances), 2)
+
+    def test_to_xml_team_filter_list(self):
+        xml_str = self.ed.to_video_analysis_xml(team_id=[1, 2], tag_period_starts=False)
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
+        self.assertEqual(len(instances), 4)
+
+    def test_to_xml_player_filter(self):
+        xml_str = self.ed.to_video_analysis_xml(player_id=20, tag_period_starts=False)
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
+        self.assertEqual(len(instances), 2)
+
+    def test_to_xml_minute_range(self):
+        xml_str = self.ed.to_video_analysis_xml(min_minute=40, max_minute=50, tag_period_starts=False)
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
+        self.assertEqual(len(instances), 2)
+
+    def test_to_xml_databallpy_events_filter(self):
+        xml_str = self.ed.to_video_analysis_xml(databallpy_events=["shot"], tag_period_starts=False)
+        root = self._parse(xml_str)
+        instances = root.findall(".//instance")
[…]

[…]
