+++ tests/schemas/test_event_data_to_xml.py
@@ -1,5 +1,5 @@
-import tempfile
 import os
+import tempfile
 import unittest
 import xml.etree.ElementTree as ET
 
@@ -8,9 +8,18 @@ import pandas as pd
 from databallpy import Event, LabelDict, events_to_xml
 from databallpy.schemas.event_data import EventData
 
+_T0 = pd.Timestamp("2023-01-14 12:00:00", tz="UTC")
+
 
 def make_event_data():
-    """Minimal 4-row EventData spanning 2 periods."""
+    """Minimal 4-row EventData spanning 2 periods.
+
+    Datetimes are anchored at _T0 (first pass in H1 = 0 s):
+      event 1 (H1 pass,  player 10): _T0          →  t =    0 s
+      event 2 (H1 shot,  player 20): _T0 + 2640 s →  t = 2640 s
+      event 3 (H2 pass,  player 10): _T0 + 3600 s →  t = 3600 s  (~60 min incl. break)
+      event 4 (H2 tackle,player 20): _T0 + 5455 s →  t = 5455 s
+    """
     data = {
         "event_id": [1, 2, 3, 4],
         "databallpy_event": ["pass", "shot", "pass", None],
@@ -24,7 +33,12 @@ def make_event_data():
         "is_successful": pd.array([True, False, True, None], dtype=pd.BooleanDtype()),
         "start_x": [0.0, 10.0, -5.0, 20.0],
         "start_y": [0.0, 5.0, -3.0, 10.0],
-        "datetime": [None, None, None, None],
+        "datetime": [
+            _T0,
+            _T0 + pd.Timedelta(seconds=2640),
+            _T0 + pd.Timedelta(seconds=3600),
+            _T0 + pd.Timedelta(seconds=5455),
+        ],
         "original_event_id": [100, 200, 300, 400],
         "original_event": ["pass", "shot", "pass", "tackle"],
     }
@@ -38,7 +52,7 @@ class TestEventDataToXml(unittest.TestCase):
     def _parse(self, xml_str):
         # Strip the XML declaration before parsing
         if xml_str.startswith("<?xml"):
-            xml_str = xml_str[xml_str.index("?>") + 2:].lstrip()
+            xml_str = xml_str[xml_str.index("?>") + 2 :].lstrip()
         return ET.fromstring(xml_str)
 
     def _instance_codes(self, root):
@@ -73,20 +87,26 @@ class TestEventDataToXml(unittest.TestCase):
         self.assertEqual(len(instances), 2)
 
     def test_to_xml_minute_range(self):
-        xml_str = self.ed.to_video_analysis_xml(min_minute=40, max_minute=50, tag_period_starts=False)
+        xml_str = self.ed.to_video_analysis_xml(
+            min_minute=40, max_minute=50, tag_period_starts=False
+        )
         root = self._parse(xml_str)
         instances = root.findall(".//instance")
         self.assertEqual(len(instances), 2)
 
     def test_to_xml_databallpy_events_filter(self):
-        xml_str = self.ed.to_video_analysis_xml(databallpy_events=["shot"], tag_period_starts=False)
+        xml_str = self.ed.to_video_analysis_xml(
+            databallpy_events=["shot"], tag_period_starts=False
+        )
         root = self._parse(xml_str)
         instances = root.findall(".//instance")
         self.assertEqual(len(instances), 1)
         self.assertEqual(instances[0].find("code").text, "shot")
 
     def test_to_xml_is_successful_filter(self):
-        xml_str = self.ed.to_video_analysis_xml(is_successful=True, tag_period_starts=False)
+        xml_str = self.ed.to_video_analysis_xml(
+            is_successful=True, tag_period_starts=False
+        )
         root = self._parse(xml_str)
         instances = root.findall(".//instance")
         self.assertEqual(len(instances), 2)
[…]
