+++ databallpy/schemas/event_data.py
@@ -85,18 +85,18 @@ class EventData(pd.DataFrame):
     @provider.setter
     def provider(self, _):
         raise AttributeError("Cannot set provider attribute of event data")
-       
+
     def to_video_analysis_xml(
         self,
         *,
         output: str = "string",
-        team_id: int | str | list[int | str] | None=None,
-        player_id: int | str | list[int | str] | None=None,
-        min_minute: int | None =None,
-        max_minute: int | None =None,
-        databallpy_events: list[str] | None=None,
-        original_events: list[str] | None=None,
-        is_successful: bool | None =None,
+        team_id: int | str | list[int | str] | None = None,
+        player_id: int | str | list[int | str] | None = None,
+        min_minute: int | None = None,
+        max_minute: int | None = None,
+        databallpy_events: list[str] | None = None,
+        original_events: list[str] | None = None,
+        is_successful: bool | None = None,
         before_seconds: float = 3.0,
         after_seconds: float = 3.0,
         code_column: str = "databallpy_event",
@@ -142,8 +142,8 @@ class EventData(pd.DataFrame):
             Pretty-printed XML string when ``output="string"``, otherwise the
             path to the written file.
         """
-        from databallpy.utils.to_xml import Event, LabelDict, events_to_xml
         from databallpy.utils.constants import DATABALLPY_EVENTS
+        from databallpy.utils.to_xml import Event, LabelDict, events_to_xml
 
         # --- Validation ---
         if before_seconds < 0:
@@ -163,6 +163,7 @@ class EventData(pd.DataFrame):
             unknown = [e for e in databallpy_events if e not in DATABALLPY_EVENTS]
             if unknown:
                 import warnings
+
                 warnings.warn(
                     f"Unknown databallpy_events values: {unknown}. "
                     f"Valid values are: {DATABALLPY_EVENTS}",
@@ -192,17 +193,11 @@ class EventData(pd.DataFrame):
         filtered = self[mask]
 
         # --- Time computation ---
-        abs_s = filtered["minutes"] * 60 + filtered["seconds"]
-
-        period_offsets: dict = {}
-        for pid in sorted(filtered["period_id"].unique()):
-            period_offsets[pid] = (
-                0.0 if pid == 1 else float(abs_s[filtered["period_id"] == pid].min())
-            )
-
-        relative_t = abs_s - filtered["period_id"].map(period_offsets)
-        start_t = (relative_t - before_seconds).clip(lower=0.0)
-        end_t = relative_t + after_seconds
+        # All times are absolute match seconds so that H2 follows H1 on the
+        start_datetime = self.loc[
+            (self["period_id"] == 1) & (self["databallpy_event"].isin(["pass", "shot"])),
+            "datetime",
+        ].min()
 
         # --- Build events dict ---
         events_dict: dict = {}
@@ -210,17 +205,35 @@ class EventData(pd.DataFrame):
         # Period start markers
         if tag_period_starts:
             period_code_map = {1: "1H", 2: "2H", 3: "ET1", 4: "ET2", 5: "PK"}
-            for pid in sorted(period_offsets.keys()):
+            for pid in sorted(period_code_map.keys()):
                 code = period_code_map.get(pid, f"ET{pid - 2}")
+                if pid not in self["period_id"].to_list():
+                    continue
+                period_start_datetime = self.loc[
+                    (self["period_id"] == pid)
+                    & (self["databallpy_event"].isin(["pass", "shot"])),
+                    "datetime",
+                ].min()
+                pid_t = (period_start_datetime - start_datetime).total_seconds()
                 events_dict[f"period_start_{pid}"] = Event(
                     id=f"p{pid}",
                     code=code,
[…]
