### 2026-03-17 #10223 @editorialbot
[REVIEW]: DataBallPy: Load, Synchronise, and Analyse your Soccer Data
**Submitting author:** @Alek050 (<a href="http://orcid.org/0000-0003-4056-7274">Alexander Oonk</a>)
**Repository:** https://github.com/Alek050/databallpy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.7.3
