### 2026-03-18 faf7b70
Add test for get_game with FIFA event data provider
