### 2026-03-23 f26b9ae
removed bs4 as dependency
