+++ databallpy/schemas/event_data.py
@@ -54,6 +54,7 @@ class EventDataSchema(pa.DataFrameModel):
     )
     to_player_name: Optional[pa.typing.Series[str]] = pa.Field(nullable=True)
     event_type_id: Optional[pa.typing.Series[int]] = pa.Field(ge=-1)
+    team_name: Optional[pa.typing.Series[str]] = pa.Field(nullable=True)
 
 
 class EventData(pd.DataFrame):
@@ -84,3 +85,203 @@ class EventData(pd.DataFrame):
     @provider.setter
     def provider(self, _):
         raise AttributeError("Cannot set provider attribute of event data")
+
+    def to_video_analysis_xml(
+        self,
+        *,
+        output: str = "string",
+        team_id: int | str | list[int | str] | None = None,
+        player_id: int | str | list[int | str] | None = None,
+        min_minute: int | None = None,
+        max_minute: int | None = None,
+        databallpy_events: list[str] | None = None,
+        original_events: list[str] | None = None,
+        is_successful: bool | None = None,
+        before_seconds: float = 3.0,
+        after_seconds: float = 3.0,
+        code_column: str = "databallpy_event",
+        tag_period_starts: bool = True,
+        time_decimals: int = 2,
+    ) -> str:
+        """Export filtered events to XML compatible with SportsCode/Longomatch.
+
+        Parameters
+        ----------
+        output : str
+            If ``"string"`` (default), return the XML as a string. Otherwise,
+            treat the value as a file path, write the XML there, and return the
+            path as a string.
+        team_id : scalar or list, optional
+            Filter to specific team(s).
+        player_id : scalar or list, optional
+            Filter to specific player(s).
+        min_minute : int, optional
+            Include only events at or after this minute.
+        max_minute : int, optional
+            Include only events at or before this minute.
+        databallpy_events : list[str], optional
+            Filter to specific databallpy event types.
+        original_events : list[str], optional
+            Filter to specific original event types.
+        is_successful : bool, optional
+            Filter by success status.
+        before_seconds : float
+            Seconds before the event to start the clip (default 3.0).
+        after_seconds : float
+            Seconds after the event to end the clip (default 3.0).
+        code_column : str
+            Column to use as the event code (default "databallpy_event").
+        tag_period_starts : bool
+            Whether to add period start markers (default True).
+        time_decimals : int
+            Decimal places for time values (default 2).
+
+        Returns
+        -------
+        str
+            Pretty-printed XML string when ``output="string"``, otherwise the
+            path to the written file.
+        """
+        from databallpy.utils.constants import DATABALLPY_EVENTS
+        from databallpy.utils.to_xml import Event, LabelDict, events_to_xml
+
+        # --- Validation ---
+        if before_seconds < 0:
+            raise ValueError(f"before_seconds must be >= 0, got {before_seconds}")
+        if after_seconds <= 0:
+            raise ValueError(f"after_seconds must be > 0, got {after_seconds}")
+        if code_column not in self.columns:
+            raise ValueError(
+                f"code_column '{code_column}' is not a column in the DataFrame. "
+                f"Available columns: {list(self.columns)}"
+            )
+        if min_minute is not None and max_minute is not None and min_minute > max_minute:
+            raise ValueError(
+                f"min_minute ({min_minute}) must be <= max_minute ({max_minute})"
+            )
+        if databallpy_events is not None:
+            unknown = [e for e in databallpy_events if e not in DATABALLPY_EVENTS]
+            if unknown:
+                import warnings
+
+                warnings.warn(
+                    f"Unknown databallpy_events values: {unknown}. "
+                    f"Valid values are: {DATABALLPY_EVENTS}",
+                    UserWarning,
[…]
