### 2026-03-25 3156ba4
Second implementation spearman pitch control
The calculation of the pitch control was changed to be more true to the original intent of spearman.
