+++ databallpy/features/pitch_control_spearman.py
@@ -82,7 +82,8 @@ def build_ndarray(
     return np.asarray(output)
 
 def prepare_data(
-        data: pd.Series
+        data: pd.Series,
+        ball: bool = False
         ) -> tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
     """
     Function to prepare all needed data from a single frame
@@ -108,8 +109,6 @@ def prepare_data(
     Velocities of the away players in the form of an array with the dimensions (n, 2)
     """
 
-    ball_pos = build_ndarray(data)
-
     pattern_home_pos = re.compile(rf"home_(\d+)_(x|y)$")
     home_pos = build_ndarray(data, pattern_home_pos, "position")
 
@@ -122,7 +121,11 @@ def prepare_data(
     pattern_away_vel = re.compile(rf"away_(\d+)_(vx|vy)$")
     away_vel = build_ndarray(data, pattern_away_vel, "velocity")
 
-    return home_pos, away_pos, home_vel, away_vel, ball_pos
+    if ball:
+        ball_pos = build_ndarray(data)
+        return home_pos, away_pos, home_vel, away_vel, ball_pos
+    
+    return home_pos, away_pos, home_vel, away_vel
 
 def time_to_intercept(
     p_players: np.ndarray,
@@ -166,27 +169,15 @@ def time_to_intercept(
         to get to the ball.
     """
 
-    v = (
-        p_ball[:, None, :] - p_players[None, :, :]
-    )  # Relative motion vector between Pressing Players and Players Under Pressure
-
-    u_mag = np.linalg.norm(p_players, axis=-1)  # velocitie of Pressing Players velocity
-    v_mag = np.linalg.norm(v, axis=-1)  # velocitie of relative motion vector
-    dot_product = np.sum(p_players * v, axis=-1)
-
-    epsilon = 1e-10  # We add epsilon to avoid dividing by zero (which throws a warning)
-    angle = np.arccos(dot_product / (u_mag * v_mag + epsilon))
-
     r_reaction = (
         p_players + v_players * reaction_time
     )  # Adjusted position of Pressing Players after reaction time
     d = p_ball[:, None, :] - r_reaction[None, :, :]  # Distance vector after reaction time
 
     t = (
-        u_mag * angle / np.pi  # Time contribution from angular adjustment
-        + reaction_time  # Add reaction time
+        reaction_time
         + np.linalg.norm(d, axis=-1) / max_velocity
-    )  # Time contribution from running
+    )  
 
     return t
 
@@ -240,16 +231,30 @@ def compute_team_tti(
         to get to the ball.
     """
     
-    time_home_players = time_to_intercept(home_player_position, ball_position, home_player_velocity, reaction_time, max_velocity)
-    time_away_players = time_to_intercept(away_player_position, ball_position, away_player_velocity, reaction_time, max_velocity)
+    t_home = time_to_intercept(
+        home_player_position, ball_position,
+        home_player_velocity, reaction_time, max_velocity
+    )[0]
 
-    return time_home_players, time_away_players
+    t_away = time_to_intercept(
+        away_player_position, ball_position,
+        away_player_velocity, reaction_time, max_velocity
+    )[0]
+
+    times = np.concatenate([t_home, t_away])
+    teams = np.array([1] * len(t_home) + [-1] * len(t_away))
+
+    sort_idx = np.argsort(times)
+    teams = teams[sort_idx]
+
+    t_diff = np.diff(times, append= times[-1])
+
+    return teams, t_diff
 
 
 def probability_to_intercept(
     time_to_intercept: np.ndarray,
     tti_sigma: float, 
-    tti_time_threshold: float
 ) -> np.ndarray:
     
     """
@@ -272,57 +277,13 @@ def probability_to_intercept(
[…]
