### 2026-03-25 Alek050/databallpy#356 @melund
JOSS Review: PyPI release v0.7.2 bug with get_open_game()
The released version on Pypi (0.7.2) seems to fail when calling `get_open_game()` with the error `ValueError: Unknown type of tracab metadata`. I tried the develop branch where it seemed to work.

Since the `get_open_game()` function is the main entry for new users without their own data and it is among the first code examples in the docs I would suggest making a new patch release before the JOSS paper is accepted. Just so published version matches what we reviewed.

https://github.com/openjournals/joss-reviews/issues/10223
