### 2026-03-25 Alek050/databallpy#357 @melund
JOSS Review: API documentation gaps
the auto-generated API docs cover the `Game` class and vizsualization functions. But not other things like data parsers, event types (ShotEvent, PassEvent etc. ) or utility funcs. A new user who wants to work with the event objects would need to read the source code. I would suggest expanding the API docs to cover the public classes and tracking data schema methods like `get_covered_distance` etc. Since these are the main ways a user interact with the package after loading data.

It don't think it ha be very exhaustive. Just the methods and shown in the examples would be fine.

This also not a blocker for the review, but more a suggestion which stumbled on.

https://github.com/openjournals/joss-reviews/issues/10223
