### 2026-03-25 Alek050/databallpy#360 @melund
JOSS reivew: CONTRIBUTING.md at the repo root
The contribution guidelines are in `docs/contributing` and seem very well written. But there is no `CONTRIBUTING.md` file a the repo root. I think the GitHub UI will surface the root level `CONTRIBUTING.md` file when users open new issues or PR.

So adding one (even a short file with links to the full docs) will make the guide lines easier to find.

https://github.com/openjournals/joss-reviews/issues/10223
