### 2026-03-25 Alek050/databallpy#361 @melund
JOSS Reivew - Minor. Caches cause issues when switching between versions
When I switched between different PyPi version and versions install from source I ran into some issues with the cached game data. Especially I got some `FileNotFound`erros.

I didn't really investigate further, but if there are some incompatiblities between versions, then maybe adding a try/except around the cache functions which just re-downloads on failure would make the experience better.

This is really just a suggestion.

https://github.com/openjournals/joss-reviews/issues/10223
