+++ joss-paper/paper.bib
@@ -17,7 +17,7 @@
 @article{Anzer2021,
    abstract = {Due to the low scoring nature of football (soccer), shots are often used as a proxy to evaluate team and player performances. However, not all shots are created equally and their quality differs significantly depending on the situation. The aim of this study is to objectively quantify the quality of any given shot by introducing a so-called expected goals (xG) model. This model is validated statistically and with professional match analysts. The best performing model uses an extreme gradient boosting algorithm and is based on hand-crafted features from synchronized positional and event data of 105, 627 shots in the German Bundesliga. With a ranked probability score (RPS) of 0.197, it is more accurate than any previously published expected goals model. This approach allows us to assess team and player performances far more accurately than is possible with traditional metrics by focusing on process rather than results.},
    author = {Gabriel Anzer and Pascal Bauer},
-   doi = {https://doi.org/10.3389/fspor.2021.624475},
+   doi = {10.3389/fspor.2021.624475},
    issn = {26249367},
    journal = {Frontiers in Sports and Active Living},
    keywords = {XG,applied machine learning,event data,expected goals,football,positional data,soccer,sports analytics},
@@ -376,6 +376,7 @@ journal = {Journal of Open Source Software}
 @article{sotudeh2025,
   title={The principles of tactical formation identification in association football (soccer)—a survey},
   author={Sotudeh, Hadi},
+  doi={10.3389/fspor.2024.1512386},
   journal={Frontiers in Sports and Active Living},
   volume={6},
   pages={1512386},
