### 2026-04-01 9905424
try except in get saved game when cache is true
