### 2026-04-01 a3f1419
Added try except on cache fail for open game
