### 2026-04-08 Alek050/databallpy#365 @felixchenier
As a complementary issue, it seems that the documentation may not be up to date at some places, at least here where the default (non-working) value is provider="sportec" while the parameters section indicates "dfl".

https://databallpy.readthedocs.io/en/latest/api_docs/game_api.html#databallpy.get_open_game

```

get_open_game(provider: str = 'sportec', game_id: str = 'J03WMX', verbose: bool = True, use_cache: bool = True) → Game
[[source]](https://databallpy.readthedocs.io/en/latest/_modules/databallpy/utils/get_game.html#get_open_game)

    Function to load a game object from an open datasource

    Parameters:

            provider (str, optional) – What provider to get the open data from. Defaults to “dfl”. Options are [“metrica”, “dfl”, “sportec”, “tracab”]

            game_id (str, optional) – The Game id of the open game. Defaults to ‘J03WMX’,

            verbose (bool, optional) – Whether or not to print info about progress

            terminal (in the)

            True. (Defaults to)

            use_cache (bool, optional) – Use cached version of match if available.

    Returns:

        All information about the game
    Return type:

        Game

```
