### 2026-04-08 Alek050/databallpy#365 @felixchenier
https://github.com/openjournals/joss-reviews/issues/10223
