### 2026-04-08 Alek050/databallpy#368 @felixchenier
[JOSS Review] PyPI stats
On your manuscript, for this text:

DataBallPy has shown to be increasingly used by coders, practitioners, and researchers. The package has been downloaded over 47.000 times on PyPI, averaging more than 250 downloads per week.

I do not think average downloads on PyPI is a measure of popularity for a project, since these downloads count are highly biased: https://packaging.python.org/en/latest/guides/analyzing-pypi-package-downloads/

Consequently, PyPI stopped displaying download statistics. I would suggest removing this sentence from the manuscript since it does not scientifically show if the software is good or bad, or even popular or not.

https://github.com/openjournals/joss-reviews/issues/10223
