### 2026-04-08 #10223 @felixchenier
I'm done for now; I'll continue the review once this issue has been solved because in the mean time I could not check the missing checkboxes.

https://github.com/Alek050/databallpy/issues/365
