### 2026-04-14 73ee1a0
Remove bs4 as dependency (#355)
* removed bs4 as dependency

* linters

* updated lock file
