+++ databallpy/data_parsers/event_data_parsers/sportec_parser.py
@@ -1,7 +1,7 @@
 import os
+import xml.etree.ElementTree as ET
 from pathlib import Path
 
-import bs4
 import chardet
 import numpy as np
 import pandas as pd
@@ -155,9 +155,9 @@ def _get_sportec_event_data(
         encoding = chardet.detect(f.read())["encoding"]
     with open(event_data_loc, "r", encoding=encoding) as file:
         lines = file.read()
-    soup = bs4.BeautifulSoup(lines, "xml")
+    root = ET.fromstring(lines)
 
-    all_events = soup.find_all("Event", {"X-Position": True})
+    all_events = root.findall(".//Event[@X-Position]")
 
     def update_results_dict(res_dict, i, **kwargs):
         for key in res_dict.keys():
@@ -182,7 +182,7 @@ def _get_sportec_event_data(
     }
 
     pitch_center, period_start_times, swap_half = _initialize_search_variables(
-        soup, metadata.home_team_id
+        root, metadata.home_team_id
     )
     metadata.periods_changed_playing_direction = [swap_half]
 
@@ -196,10 +196,10 @@ def _get_sportec_event_data(
     for idx, event in enumerate(all_events):
         kwargs = {}
         kwargs["set_piece"] = SPORTEC_SET_PIECES_MAP.get(
-            event.find_next().name, "no_set_piece"
+            next(iter(event)).tag, "no_set_piece"
         )
 
-        kwargs["datetime"] = pd.to_datetime(event["EventTime"]).tz_convert(
+        kwargs["datetime"] = pd.to_datetime(event.get("EventTime")).tz_convert(
             "Europe/Berlin"
         )
         dt_idx = int(kwargs["datetime"] >= period_start_times[1])
@@ -210,32 +210,31 @@ def _get_sportec_event_data(
         kwargs["minutes"] = int((45 * dt_idx) + time_diff_s // 60)
         kwargs["seconds"] = time_diff_s % 60
         kwargs["event_id"] = idx
-        kwargs["original_event_id"] = int(event["EventId"])
+        kwargs["original_event_id"] = int(event.get("EventId"))
 
-        kwargs["start_x"] = float(event["X-Position"]) - pitch_center[0]
-        kwargs["start_y"] = float(event["Y-Position"]) - pitch_center[1]
+        kwargs["start_x"] = float(event.get("X-Position")) - pitch_center[0]
+        kwargs["start_y"] = float(event.get("Y-Position")) - pitch_center[1]
 
         if kwargs["period_id"] == swap_half:
             kwargs["start_x"] *= -1
             kwargs["start_y"] *= -1
 
-        next_tag = event
-        while next_tag is not None and next_tag.name not in ALL_SPORTEC_EVENTS:
-            next_tag = next_tag.find_next()
+        event = next(
+            (e for e in event.iter() if e is not event and e.tag in ALL_SPORTEC_EVENTS),
+            next(iter(event)),
+        )
 
-        event = next_tag if next_tag is not None else event.find_next()
-
-        kwargs["original_event"] = event.name
+        kwargs["original_event"] = event.tag
         kwargs["player_id"] = event.get("Player", event.get("Winner"))
         kwargs["team_id"] = event.get("Team", event.get("WinnerTeam"))
 
-        if event.name == "ShotAtGoal":
+        if event.tag == "ShotAtGoal":
             kwargs, shot_event = _handle_shot_event(event, metadata, kwargs)
             databallpy_events["shot_events"][shot_event.event_id] = shot_event
-        elif event.name == "Play":
+        elif event.tag == "Play":
             kwargs, pass_event = _handle_play_event(event, metadata, kwargs)
[…]
