+++ docs/api_docs/tracking_data_api.rst
@@ -1,0 +1,12 @@
+Tracking Data API
+=================
+
+TrackingData
+---------------------------------
+`databallpy.schemas.TrackingData`
+
+.. autoclass:: databallpy.schemas.TrackingData
+    :members:
+    :show-inheritance:
+    :inherited-members:
+    :no-index:
