### 2026-04-14 Alek050/databallpy#368 @Alek050
Thanks for this reference, I was not aware of this. I removed it from the manuscript.
