### 2026-04-14 a46d223
Merged latest version of develop and removed pypi stats from paper.md (#369)
* fix for pandera & pandas integration python 3.14 (#348)

* Enh to xml (#353)

* Added functionality to export to xml

* added docs

* Added doc files

* Added try except on cache fail for open game

* Added CONTRIBUTING.md file (#364)

* Remove bs4 as dependency (#355)

* removed bs4 as dependency

* linters

* updated lock file

* Removed pypi download stats from paper.md
