+++ joss-paper/paper.md
@@ -125,7 +125,7 @@ plt.show()
 
 # Research impact statement
 
-`DataBallPy` has shown to be increasingly used by coders, practitioners, and researchers. The package has been downloaded over 47.000 times on PyPI, averaging more than 250 downloads per week. The project has over 60 GitHub stars. Issues and PR's are being opened by users outside of the network of the original owners and maintainers. On top of that, `DataBallPy` has been mentioned in numerous published scientific papers [@Robertson2023; @Anzer2025; @Zhang2025; @Oonk2025a]. Moreover, the largest currently open-sourced dataset of tracking and event data showcased how `DataBallPy` can be used to synchronise the two sources [@Bassek2025]. Lastly, authors who introduce new metrics propose to open a PR with their metric so it is easily available for the scientific community [@Bischofberger2025]. Together, this shows that `DataBallPy` has a wide range of users and the package is growing outside of the reach of the original owners and maintainers.
+`DataBallPy` has shown to be increasingly used by coders, practitioners, and researchers. The project has over 60 GitHub stars. Issues and PR's are being opened by users outside of the network of the original owners and maintainers. On top of that, `DataBallPy` has been mentioned in numerous published scientific papers [@Robertson2023; @Anzer2025; @Zhang2025; @Oonk2025a]. Moreover, the largest currently open-sourced dataset of tracking and event data showcased how `DataBallPy` can be used to synchronise the two sources [@Bassek2025]. Lastly, authors who introduce new metrics propose to open a PR with their metric so it is easily available for the scientific community [@Bischofberger2025]. Together, this shows that `DataBallPy` has a wide range of users and the package is growing outside of the reach of the original owners and maintainers.
 
 # AI usage disclosure
 
