### 2026-04-14 Alek050/databallpy#365 @Alek050
Hi @felixchenier,

To reply to the raised problems here:
1. Caching behaviour changed in version 0.7.2. Another suggestion was raised by @melund which should be fixed in #361. I will release a new version soon to make sure this will work again.
2. The `KeyError` is raised not because `DataBallPy` did not download/ read in the data, but because the tracking and event data was not synchronised yet. I am assuming you used the code snippet available in the [visualisation docs](https://databallpy.readthedocs.io/en/latest/getting_started/visualizations_page.html) at the end of the page. The key `'event_id'` is only added ones the tracking and even data is synchronised using `game.synchronise_tracking_and_event_data()`. I will update the docs to make this more clear.
3. Fair point, in practice `"sportec"` and `"dfl"` produce the exact same behaviour, but it is unclear for users. I will change this as well.
