### 2026-04-14 0175b66
Updated get open game docstrings and added synchronisation to visualisation documentation page
