### 2026-04-14 Alek050/databallpy#357 @Alek050
Hi @melund,

Thanks for the suggestion. I added auto-api docs for tracking data and event data including its methods. I also added them for the loading in different datasources.

I choose not to add docs for `ShotEvent` and other events since I plan to remove it from from the package in the future. Although the functionalities will be available in other places of the package, I'd rather not have too many users rely on it because of that for now.
