+++ databallpy/schemas/tracking_data.py
@@ -273,27 +273,30 @@ class TrackingData(pd.DataFrame):
         allow_overwrite: bool = False,
     ) -> None:
         """Function that adds acceleration columns to the tracking data based on the
-           position columns
+        position columns.
 
         Args:
-            self,
-            column_ids (str | list[str]): columns for which acceleration should be calculated
-            filter_type (str, optional): filter type to use. Defaults to None. Options are `moving_average` and `savitzky_golay`.
-            window_length (int, optional): window size for the filter. Defaults to 25.
-            polyorder (int, optional): polynomial order for the filter. Defaults to 2.
-            max_acceleration (float, optional): maximum value for the acceleration. Defaults to np.inf.
-            allow_overwrite (bool): Whether or not it is allowed to overwrite existing valuesNote: if "_acceleration" exists, but "_ax" and "_ay" not, and allow_overwrite is set to False, "_ax" and "_ay" will be computed and added, but "_acceleration" is kept the same, and therefore does not correspond with the other values. Defaults to False.
+            column_ids (str | list[str]): Columns for which acceleration should be calculated.
+            filter_type (str, optional): Filter type to use. Defaults to None.
+                Options are `moving_average` and `savitzky_golay`.
+            window_length (int, optional): Window size for the filter. Defaults to 25.
+            polyorder (int, optional): Polynomial order for the filter. Defaults to 2.
+            max_acceleration (float, optional): Maximum value for the acceleration. Defaults to np.inf.
+            allow_overwrite (bool): Whether or not it is allowed to overwrite existing values.
 
         Returns:
             None
 
         Raises:
-            ValueError: if filter_type is not one of `moving_average`, `savitzky_golay`, or None.
-            ValueError: if velocity was not found in the DataFrame for the input_columns.
+            ValueError: If filter_type is not one of `moving_average`, `savitzky_golay`, or None.
+            ValueError: If velocity was not found in the DataFrame for the input_columns.
 
         Note:
-            The function will delete the columns in input_columns with the acceleration if
-            they already exist.
+            If "_acceleration" exists, but "_ax" and "_ay" do not, and allow_overwrite is False,
+            "_ax" and "_ay" will be computed and added, but "_acceleration" is kept unchanged.
+            Therefore, it may not correspond with the other values.
+
+            The function will delete acceleration columns if they already exist.
         """
 
         if isinstance(column_ids, str):
