+++ docs/api_docs/dataparsers_api.rst
@@ -1,0 +1,30 @@
+Data Parsers API
+================
+
+Event Data Parsers
+------------------
+
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_instat_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_metrica_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_metrica_open_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_scisports_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_sportec_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_sportec_open_event_data
+.. autofunction:: databallpy.data_parsers.event_data_parsers.load_statsbomb_event_data
+
+
+Tracking Data Parsers
+---------------------
+
+.. autofunction:: databallpy.data_parsers.tracking_data_parsers.load_inmotio_tracking_data
+.. autofunction:: databallpy.data_parsers.tracking_data_parsers.load_metrica_open_tracking_data
+.. autofunction:: databallpy.data_parsers.tracking_data_parsers.load_metrica_tracking_data
+.. autofunction:: databallpy.data_parsers.tracking_data_parsers.load_sportec_open_tracking_data
+.. autofunction:: databallpy.data_parsers.tracking_data_parsers.load_tracab_tracking_data
+
+
+Parse from Kloppy
+-----------------
+
+.. autofunction:: databallpy.data_parsers.kloppy_parsers.convert_kloppy_tracking_dataset
+.. autofunction:: databallpy.data_parsers.kloppy_parsers.convert_kloppy_event_dataset
