+++ docs/api_docs/tracking_data_api.rst
@@ -8,5 +8,4 @@ TrackingData
 .. autoclass:: databallpy.schemas.TrackingData
     :members:
     :show-inheritance:
-    :inherited-members:
     :no-index:
