### 2026-04-14 57d9465
More elaborate API docs (#373)
* WIP

* Updated auto api docs
