+++ docs/conf.py
@@ -8,7 +8,7 @@ author = 'Alexander Oonk'
 comments_config = {'hypothesis': False, 'utterances': False}
 copyright = 'Copyright 2024, Alexander Oonk'
 exclude_patterns = ['**.ipynb_checkpoints', '.DS_Store', '.env', 'Thumbs.db', '_build']
-extensions = ['sphinx_togglebutton', 'sphinx_copybutton', 'myst_nb', 'jupyter_book', 'sphinx_thebe', 'sphinx_comments', 'sphinx_external_toc', 'sphinx.ext.intersphinx', 'sphinx_design', 'sphinx_book_theme', 'sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.viewcode', 'sphinx_jupyterbook_latex', 'sphinx_multitoc_numbering']
+extensions = ['sphinx_togglebutton', 'sphinx_copybutton', 'myst_nb', 'jupyter_book', 'sphinx_thebe', 'sphinx_comments', 'sphinx_external_toc', 'sphinx.ext.intersphinx', 'sphinx_design', 'sphinx_book_theme', 'sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.viewcode', 'sphinx_jupyterbook_latex', 'sphinx_multitoc_numbering', "matplotlib.sphinxext.plot_directive",]
 external_toc_exclude_missing = False
 external_toc_path = '_toc.yml'
 html_baseurl = ''
