### 2026-04-14 cf7cce1
Fifa parser (#374)
* Update databallpy code with FIFA data support

* fixed linters

* continue fifa implementation

* WIP

* small changes

* Added tests

* linters

* fixed bug

* fixed bug

* Update FIFA changelog

* linters and tests

* Add test for get_game with FIFA event data provider

---------

Co-authored-by: ouyang1030 <yangstal1023@163.com>
