+++ tests/test_databallpy.py
@@ -2,4 +2,4 @@ from databallpy import __version__
 
 
 def test_version():
-    assert __version__ == "0.7.2"
+    assert __version__ == "0.7.3"
