### 2026-04-14 Alek050/databallpy#356 @Alek050
Thanks, I just released 0.7.3 which should fix this.
