### 2026-04-14 #10223 @Alek050
Hi @melund and @felixchenier,

Thanks a lot for the valueable feedback. I think I have adressed most of the opened issues and concers you've raised. I just released version 0.7.3 which includes fixes to these points.

I tried to reproduce the error raised in https://github.com/Alek050/databallpy/issues/365, but could not do so. I hope this is magically fixed in the new release as well, but otherwise I need to dive into tis further.

If there is anything else I can do / improve do not hesitate to reach out!
