### 2026-04-20 #10223 @Alek050
@arfon

I've addeda all requested changes in the latest version (0.7.3) and created a doi on zenodo: https://doi.org/10.5281/zenodo.19661454.

Let me know If you are expecting anythinge more from me.
