### 2026-04-21 2a81eee
Fix author field from 'given-name' to 'given-names'
