### 2026-04-21 #10223 @arfon
@Alek050 – could you please merge this PR which should fix the author names: https://github.com/Alek050/databallpy/pull/378
